move files to destination layer
