+++ mcr-core/mcr_meeting/app/domain/docx_generation/templated_docx_generator.py
