+++ mcr-core/mcr_meeting/app/domain/report_content/template_renderer.py
