+++ mcr-core/mcr_meeting/app/domain/report_content/templates/detailed_synthesis.md.jinja
