+++ mcr-core/mcr_meeting/app/services/transcription_task_service.py
@@ -3,9 +3,11 @@ from io import BytesIO
 from mcr_meeting.app.exceptions.exceptions import (
     NotFoundException,
 )
-from mcr_meeting.app.infrastructure.s3 import get_transcription_object_name
+from mcr_meeting.app.infrastructure.s3 import (
+    get_file_from_s3,
+    get_transcription_object_name,
+)
 from mcr_meeting.app.models import Meeting
-from mcr_meeting.app.services.s3_service import get_file_from_s3
 
 
 def get_formatted_transcription_from_s3(meeting: Meeting) -> BytesIO:
