+++ mcr-core/mcr_meeting/app/use_cases/evaluate_transcription_from_s3.py
@@ -4,7 +4,6 @@ from mcr_meeting.app.infrastructure.celery import enqueue_evaluation_from_s3_tas
 
 
 def evaluate_transcription_from_s3(zip_name: str) -> None:
-
     if not is_zip_filename(zip_name):
         raise BadRequestException("zip_name must reference a .zip file.")
     enqueue_evaluation_from_s3_task(zip_name)
